+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `spotlight` handling of 5-way events after the pointer hides
+
 ## [2.0.0-beta.2] - 2018-05-07
 
 ### Fixed