+++ packages/spotlight/CHANGELOG.md
@@ -6,8 +6,14 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
-- `spotlight/Spottable` to not set its focused state when disabled
+- `spotlight` to retry setting focus when the window is activated
+- `spotlight` handling of 5-way events after the pointer hides
 
+## [2.0.0-beta.2] - 2018-05-07
+
+### Fixed
+
+- `spotlight/Spottable` to not add a focused state when a component had already been set as disabled
 
 ## [2.0.0-beta.1] - 2018-04-29